+++ NEWS.rst
@@ -1,3 +1,5 @@
+- [unicodedata] Update RTL_SCRIPTS for Unicode 11 and 12.
+
 4.2.3 (released 2020-01-07)
 ---------------------------